:hammer: Provide clipeveryLogger to support more flexible log configuration
