+++ composeApp/src/commonMain/kotlin/com/clipevery/log/Logger.kt
@@ -1,3 +1,9 @@
 package com.clipevery.log
 
-expect fun initLogger(logPath: String)
+expect fun initLogger(logPath: String): ClipeveryLogger
+
+interface ClipeveryLogger {
+    val logLevel: String
+    val logPath: String
+    val loggerDebugPackages: String?
+}
