+++ composeApp/src/desktopMain/kotlin/com/clipevery/net/DesktopClipClient.kt
@@ -3,6 +3,7 @@ package com.clipevery.net
 import com.clipevery.app.AppInfo
 import com.clipevery.net.plugin.SignalClientEncryption
 import com.clipevery.utils.DesktopJsonUtils
+import io.github.oshai.kotlinlogging.KotlinLogging
 import io.ktor.client.*
 import io.ktor.client.engine.cio.*
 import io.ktor.client.plugins.*
@@ -22,12 +23,21 @@ import java.io.File
 
 class DesktopClipClient(private val appInfo: AppInfo) : ClipClient {
 
+    private val clientLogger = KotlinLogging.logger {}
+
     private val client: HttpClient =
         HttpClient(CIO) {
             install(HttpTimeout) {
                 requestTimeoutMillis = 1000
             }
-            install(Logging)
+            install(Logging, configure = {
+                logger =
+                    object : Logger {
+                        override fun log(message: String) {
+                            clientLogger.debug { message }
+                        }
+                    }
+            })
             install(ContentNegotiation) {
                 json(DesktopJsonUtils.JSON, ContentType.Application.Json)
             }
